Updated Station Image handling to check for file size and image media type. Also renamed for simplicity and used the newly renamed stationimage table.

+++ RWICPreceiverApp/Providers/BlobStorageUploadProvider.cs
@@ -28,37 +28,48 @@ namespace RWICPreceiverApp.Providers
             foreach (var fileData in FileData)
             {
                 if (!string.IsNullOrEmpty(fileData.Headers.ContentDisposition.FileName))
-                {
-                    // Sometimes the filename has a leading and trailing double-quote character
-                    // when uploaded, so we trim it; otherwise, we get an illegal character exception
-                    var fileName = Path.GetFileName(fileData.Headers.ContentDisposition.FileName.Trim('"'));
-
-                    // Retrieve reference to a blob
-                    var blobContainer = BlobHelper.GetBlobContainer();
-                    var blob = blobContainer.GetBlockBlobReference(fileName);
-
-                    // Set the blob content type
-                    blob.Properties.ContentType = fileData.Headers.ContentType.MediaType;
-
-                    // Upload file into blob storage, basically copying it from local disk into Azure
-                    using (var fs = File.OpenRead(fileData.LocalFileName))
+                {                   
+                    string mediaType = fileData.Headers.ContentType.MediaType.ToString();
+                    //If the file an image media type
+                    if (System.Text.RegularExpressions.Regex.IsMatch(mediaType, "image/\\S+") )
                     {
-                        blob.UploadFromStream(fs);
+                        // Sometimes the filename has a leading and trailing double-quote character
+                        // when uploaded, so we trim it; otherwise, we get an illegal character exception
+                        var fileName = Path.GetFileName(fileData.Headers.ContentDisposition.FileName.Trim('"'));
+
+                        // Retrieve reference to a blob
+                        var blobContainer = BlobHelper.GetBlobContainer();
+                        var blob = blobContainer.GetBlockBlobReference(fileName);
+                        //if(blob.Metadata)
+                        // Set the blob content type
+                        blob.Properties.ContentType = fileData.Headers.ContentType.MediaType;                    
+
+                        // Upload file into blob storage, basically copying it from local disk into Azure
+                        using (var fs = File.OpenRead(fileData.LocalFileName))
+                        {
+                            long fileSizeInKB = (long)(fs.Length / 1024);
+                            //If the image is greater than 1 MB don't save it 
+                            if (fileSizeInKB > 1001)
+                            {
+                                continue;
+                            }
+                            blob.UploadFromStream(fs);
+                        }
+
+                        // Delete local file from disk
+                        File.Delete(fileData.LocalFileName);
+
+                        // Create blob upload model with properties from blob info
+                        var blobUpload = new BlobUploadModel
+                        {
+                            FileName = blob.Name,
+                            FileUrl = blob.Uri.AbsoluteUri,
+                            FileSizeInBytes = blob.Properties.Length
+                        };
+
+                        // Add uploaded blob to the list
+                        Uploads.Add(blobUpload);
                     }
-
-                    // Delete local file from disk
-                    File.Delete(fileData.LocalFileName);
-
-                    // Create blob upload model with properties from blob info
-                    var blobUpload = new BlobUploadModel
-                    {
-                        FileName = blob.Name,
-                        FileUrl = blob.Uri.AbsoluteUri,
-                        FileSizeInBytes = blob.Properties.Length
-                    };
-
-                    // Add uploaded blob to the list
-                    Uploads.Add(blobUpload);
                 }                
             }
